# refactor _evaluate_image to improve performance