#### Reference Issues/PRs

resolve #1009 


#### What does this implement/fix? Explain your changes.
Refactor to improve performance and readability for future fixes.

#### Any other comments?

    # before
        1    0.002    0.002  104.631  104.631 suite.py:47(run)
     2304    0.045    0.000   29.823    0.013 detection_precision_recall.py:135(_group_detections)
     7002    5.973    0.001   24.638    0.004 detection_precision_recall.py:153(_evaluate_image)

    # after
        1    0.030    0.030   90.438   90.438 suite.py:47(run)
     2304    0.045    0.000   12.626    0.005 detection_precision_recall.py:134(_group_detections)
     7002    1.980    0.000    7.281    0.001 detection_precision_recall.py:152(_evaluate_image)

`_group_detections` run time reduced from `29.823` seconds to `12.626` seconds. A total of `17.197` seconds saved or over 10% of total runtime.

---
Thanks for contributing a pull request! Please ensure you have taken a look at
the contribution guidelines: https://github.com/deepchecks/deepchecks/blob/main/CONTRIBUTING.rst